refactor: avoid using utils as naming
