+++ src/query/sql/src/planner/mod.rs
@@ -17,8 +17,8 @@ mod format;
 mod metadata;
 #[allow(clippy::module_inception)]
 mod planner;
+mod query_builders;
 mod semantic;
-mod utils;
 
 pub(crate) mod binder;
 pub mod optimizer;
@@ -33,8 +33,8 @@ pub use expression_parser::ExpressionParser;
 pub use metadata::*;
 pub use planner::Planner;
 pub use plans::ScalarExpr;
+pub use query_builders::SelectBuilder;
 pub use semantic::normalize_identifier;
 pub use semantic::validate_function_arg;
 pub use semantic::IdentifierNormalizer;
 pub use semantic::NameResolutionContext;
-pub use utils::SelectBuilder;
